+++ .storybook/manager.ts
@@ -5,7 +5,7 @@ addons.setConfig({
   theme: {
     ...themes.light,
     brandTitle: 'Ngxsmk Tel Input',
-    brandUrl: 'https://github.com/toozuuu/ngxsmk-tel-input',
+    brandUrl: 'https://github.com/NGXSMK/ngxsmk-tel-input',
     brandImage: undefined,
     brandTarget: '_self'
   }